+++ order-service/src/main/java/com/ecommerce/orderservice/OrderServiceApplication.java
@@ -2,17 +2,11 @@ package com.ecommerce.orderservice;
 
 import com.ecommerce.shared.SharedLibraryConfig;
 import com.ecommerce.shared.openapi.OpenApiConfigBase;
-import org.modelmapper.ModelMapper;
-import org.modelmapper.config.Configuration;
-import org.modelmapper.convention.MatchingStrategies;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 
-@EnableFeignClients
 @EnableDiscoveryClient
 @SpringBootApplication
 @Import({SharedLibraryConfig.class, OpenApiConfigBase.class})
@@ -22,13 +16,4 @@ public class OrderServiceApplication {
         SpringApplication.run(OrderServiceApplication.class, args);
     }
 
-    @Bean
-    public ModelMapper modelMapper() {
-        ModelMapper modelMapper = new ModelMapper();
-        modelMapper.getConfiguration()
-                .setFieldMatchingEnabled(true)
-                .setFieldAccessLevel(Configuration.AccessLevel.PRIVATE)
-                .setMatchingStrategy(MatchingStrategies.STRICT);
-        return modelMapper;
-    }
 }
