+++ order-service/src/main/java/com/ecommerce/orderservice/infrastructure/kafka/KafkaEventPublisherAdapter.java
@@ -1,0 +1,4 @@
+package com.ecommerce.orderservice.infrastructure.kafka;
+
+public class KafkaEventPublisherAdapter {
+}
