+++ order-service/src/main/java/com/ecommerce/orderservice/infrastructure/persistence/jpa/OrderJpaRepository.java
@@ -1,0 +1,13 @@
+package com.ecommerce.orderservice.infrastructure.persistence.jpa;
+
+import com.ecommerce.orderservice.infrastructure.persistence.entity.OrderEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface OrderJpaRepository extends JpaRepository<OrderEntity, UUID> {
+
+}
+
