+++ order-service/src/main/java/com/ecommerce/orderservice/interfaces/kafka/OrderEventConsumer.java
@@ -1,0 +1,4 @@
+package com.ecommerce.orderservice.interfaces.kafka;
+
+public class OrderEventConsumer {
+}
